api/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
confusable-homoglyphs==2.0.2
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.33.0
django-cors-headers==2.1.0
django-extensions==1.9.6
django-registration==2.3
django-rest-auth==0.9.2
djangorestframework==3.7.1
hashids==1.2.0
httpie==0.9.9
httpie-jwt-auth==0.3.0
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
oauthlib==2.0.6
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
